Copy files for benchmarks
