+++ .github/workflows/asv-nightly.yml
@@ -63,6 +63,12 @@ jobs:
             mv ../_results .
           fi
 
+      - name: Download catalog data
+        run: python -m ${{ env.WORKING_DIR }}/copy_benchmark_data.py
+        env:
+          ABFS_LINCCDATA_ACCOUNT_NAME: ${{ secrets.LINCC_ABFS_ACCOUNT_NAME }}
+          ABFS_LINCCDATA_ACCOUNT_KEY: ${{ secrets.LINCC_ABFS_ACCOUNT_KEY }}
+
       - name: Get nightly dates under comparison
         id: nightly-dates
         run: |
